tests/wpt/web-platform-tests/app-history/navigate-event/navigateerror-ordering-cross-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  let start_href = i.contentWindow.location.href;
  let events = [];
  i.contentWindow.appHistory.onnavigateerror = () => events.push("navigateerror");
  i.contentWindow.appHistory.onnavigate = t.step_func(e => {
    assert_equals(i.contentWindow.location.href, start_href);
    events.push("navigate");
    e.signal.onabort = () => events.push("abort");
  });
  i.contentWindow.appHistory.navigate("?1").committed.finally(t.unreached_func("navigate to ?1"));
  i.contentWindow.appHistory.navigate("?2").committed.finally(t.unreached_func("navigate to ?2"));
  assert_equals(i.contentWindow.location.search, "");
  await new Promise(resolve => i.onload = resolve);
  assert_equals(i.contentWindow.location.search, "?2");
  assert_array_equals(events, ["navigate", "abort", "navigateerror", "navigate"]);
}, "navigateerror ordering when navigate() is called repeatedly (cross-document)");
</script>
